+++ mobile/lib/features/dashboard/dashboard_page.dart
@@ -220,7 +220,10 @@ class _DashboardPageState extends State<DashboardPage> {
             ),
             const SizedBox(height: 20),
             const Center(
-              child: Text('Version 0.11', style: TextStyle(color: Colors.grey)),
+              child: Text(
+                'Version 0.11.0',
+                style: TextStyle(color: Colors.grey),
+              ),
             ),
             const SizedBox(height: 20),
           ],